Unconditionally use variants key;
Currently only functions with multiple variants use the variants table,
and functions with a single variant use an array of arguments/returns.
This means that there's less typing for single-variant functions, but
it's really annoying to change between the 2 different syntaxes.

Instead, just always be explicit about the variants and use string key
for arguments/returns tables.